+++ composer/algorithms/__init__.py
@@ -43,7 +43,6 @@ For more information about events, see :class:`~composer.core.event.Event`.
 from composer.algorithms.algorithm_hparams import AlgorithmHparams as AlgorithmHparams
 from composer.algorithms.algorithm_registry import get_algorithm_registry as get_algorithm_registry
 from composer.algorithms.algorithm_registry import list_algorithms as list_algorithms
-from composer.algorithms.alibi import AlibiHparams as AlibiHparams
 from composer.algorithms.augmix import AugMix as AugMix
 from composer.algorithms.blurpool import BlurPool as BlurPool
 from composer.algorithms.blurpool import BlurPoolHparams as BlurPoolHparams
@@ -56,6 +55,7 @@ from composer.algorithms.factorize import Factorize as Factorize
 from composer.algorithms.factorize import FactorizeHparams as FactorizeHparams
 from composer.algorithms.ghost_batchnorm import GhostBatchNorm as GhostBatchNorm
 from composer.algorithms.ghost_batchnorm import GhostBatchNormHparams as GhostBatchNormHparams
+from composer.algorithms.hparams import AlibiHparams as AlibiHparams
 from composer.algorithms.hparams import AugMixHparams as AugMixHparams
 from composer.algorithms.hparams import ColOutHparams as ColOutHparams
 from composer.algorithms.hparams import CutMixHparams as CutMixHparams
