+++ composer/algorithms/algorithm_registry.py
@@ -3,12 +3,11 @@
 from typing import Dict, List, Type
 
 from composer.algorithms.algorithm_hparams import AlgorithmHparams
-from composer.algorithms.alibi import AlibiHparams
 from composer.algorithms.blurpool import BlurPoolHparams
 from composer.algorithms.channels_last import ChannelsLastHparams
 from composer.algorithms.factorize import FactorizeHparams
 from composer.algorithms.ghost_batchnorm import GhostBatchNormHparams
-from composer.algorithms.hparams import (AugMixHparams, ColOutHparams, CutMixHparams, CutOutHparams,
+from composer.algorithms.hparams import (AlibiHparams, AugMixHparams, ColOutHparams, CutMixHparams, CutOutHparams,
                                          LabelSmoothingHparams, MixUpHparams, RandAugmentHparams)
 from composer.algorithms.layer_freezing import LayerFreezingHparams
 from composer.algorithms.no_op_model import NoOpModelHparams
